# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prospero API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Search sources by title",
        "operationId": "search_sources",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching sources",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceSearchResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      },
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Create a new source",
        "operationId": "create_source",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Source created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      }
    },
    "/api/sources/{id}": {
      "put": {
        "tags": [
          "sources"
        ],
        "summary": "Update an existing source",
        "operationId": "update_source",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Source ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Source updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Source not found"
          }
        }
      }
    },
    "/api/sources/{id}/persons": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Link a person to a source with a role",
        "operationId": "add_source_person",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Source ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkSourcePersonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Person linked to source"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      }
    },
    "/api/sources/{id}/persons/{person_id}/{role}": {
      "delete": {
        "tags": [
          "sources"
        ],
        "summary": "Remove a person from a source",
        "operationId": "remove_source_person",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Source ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "person_id",
            "in": "path",
            "description": "Person ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "path",
            "description": "Role to remove",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Person unlinked from source"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      }
    },
    "/api/sources/{id}/references": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Check if a source can be deleted (has active references)",
        "operationId": "check_source_references",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Source ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reference check result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceCheckResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSourceRequest": {
        "type": "object",
        "required": [
          "source_type",
          "title"
        ],
        "properties": {
          "doi": {
            "type": [
              "string",
              "null"
            ]
          },
          "edition": {
            "type": [
              "string",
              "null"
            ]
          },
          "isbn": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "journal_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "page_end": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "page_start": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "parent_source_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "publication_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "publisher": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          },
          "volume": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "LinkSourcePersonRequest": {
        "type": "object",
        "required": [
          "person_id",
          "role"
        ],
        "properties": {
          "person_id": {
            "type": "string"
          },
          "position": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
      "ParentSourceResponse": {
        "type": "object",
        "required": [
          "id",
          "source_type",
          "title",
          "persons"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "persons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourcePersonResponse"
            }
          },
          "publication_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "publisher": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "ReferenceCheckResponse": {
        "type": "object",
        "required": [
          "count",
          "resource_ids"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "resource_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SourcePersonResponse": {
        "type": "object",
        "required": [
          "person_id",
          "name",
          "role",
          "position"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "person_id": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "type": "string"
          },
          "sort_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "SourceResponse": {
        "type": "object",
        "required": [
          "id",
          "source_type",
          "title",
          "persons"
        ],
        "properties": {
          "doi": {
            "type": [
              "string",
              "null"
            ]
          },
          "edition": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "isbn": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "journal_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "page_end": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "page_start": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "parent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParentSourceResponse"
              }
            ]
          },
          "persons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourcePersonResponse"
            }
          },
          "publication_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "publisher": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          },
          "volume": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "SourceSearchResponse": {
        "type": "object",
        "required": [
          "id",
          "source_type",
          "title",
          "persons"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "persons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourcePersonResponse"
            }
          },
          "publication_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "source_type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "UpdateSourceRequest": {
        "type": "object",
        "properties": {
          "doi": {
            "type": [
              "string",
              "null"
            ]
          },
          "edition": {
            "type": [
              "string",
              "null"
            ]
          },
          "isbn": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "journal_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "page_end": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "page_start": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "parent_source_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "publication_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "publisher": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          },
          "volume": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    }
  }
}